backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.31.0
python-dotenv>=1.0.1
pydantic>=2.9.0
httpx>=0.27.0
aiofiles>=24.1.0
aiosqlite>=0.20.0
chromadb>=0.5.5
pyperclip>=1.9.0
litellm>=1.61.0
langgraph>=0.2.74
pytest>=8.3.0
pytest-asyncio>=0.24.0
